+++ api/src/helpers/google/README.md
@@ -1,0 +1,9 @@
+## Models/Users
+
+### jwtClient.ts
+
+JWT クライアントをサービスアカウントで認証する
+
+### insertEvent.ts
+
+イベントをカレンダーに追加する
