+++ api/src/helpers/prisma/README.md
@@ -1,0 +1,13 @@
+## Models/Users
+
+### findUser.ts
+
+メールアドレスが一致するユーザーを 1 人取り出す
+
+### registerReservation.ts
+
+メールアドレスが一致したユーザーがいた場合、予約情報のみ登録を行う
+
+### registerUserReservation.ts
+
+メールアドレスが一致したユーザーがいない場合、ユーザー情報と予約情報の登録を行う
